another nother 6

+++ 0x00-hello_world/6-size.c
@@ -6,6 +6,6 @@ printf("Size of char: %zu byte(s)\n", sizeof(char));
 printf("Size of int: %zu byte(s)\n", sizeof(int));
 printf("Size of long int: %zu byte(s)\n", sizeof(long int));
 printf("Size of long long int: %zu byte(s)\n", sizeof(double));
-printf("Size of float: %zu byte(s)\n", sizeof(long double));
+printf("Size of float: %zu byte(s)\n", sizeof(float));
 return (0);
 }
